feat: Add ser/de for HandleLogin state
This is necessary to keep a stateless server across
the two request handlers needed to implement login.

This state should be persisted to an ephemeral key/value
store, keyed on the user identifier, and with a short TTL,
to ensure the complete handshake is completed within a
reasonable time. It should also be stored encrypted for
additional security, but this is out of the scope of this feature.

- Add Rust tests
- Reflect change in JS test
- Use OPAQUE latest RFC naming conventions in JS test